Move choo route views and layouts to multiple files
