+++ web/app/views/404.js
@@ -1,0 +1,9 @@
+const html = require('choo/html')
+
+module.exports = (state, emit) => {
+  return html`
+    <div>
+      <h2>404</h2>
+    </div>
+  `
+}
