+++ web/app/views/email-confirmation.js
@@ -1,0 +1,8 @@
+const html = require('choo/html')
+
+module.exports = (state, emit) => {
+  return html`
+    <div class="flex flex-column">
+    </div>
+  `
+}
